+++ src/lib/components/chat/Suggestions.svelte
@@ -70,14 +70,6 @@
 		{$i18n.t('Suggested')}
 	{:else}
 		<!-- Keine Vorschläge -->
-
-		<div
-			class="flex w-full {$settings?.landingPageMode === 'chat'
-				? ' -mt-1'
-				: 'text-center items-center justify-center'}  self-start text-gray-600 dark:text-gray-400"
-		>
-			{$WEBUI_NAME} ‧ v{WEBUI_VERSION}
-		</div>
 	{/if}
 </div>